Refactor code according to comments
