+++ client/data/agency-dashboard/use-fetch-dashboard-sites.js
@@ -1,7 +1,6 @@
 import { useTranslate } from 'i18n-calypso';
 import { useQuery } from 'react-query';
 import { useDispatch } from 'react-redux';
-import { formatSites } from 'calypso/jetpack-cloud/sections/agency-dashboard/sites-overview/utils';
 import { wpcomJetpackLicensing as wpcomJpl } from 'calypso/lib/wp';
 import { errorNotice } from 'calypso/state/notices/actions';
 
@@ -23,7 +22,6 @@ const useFetchDashboardSites = () => {
 			} ),
 		{
 			refetchOnWindowFocus: false,
-			select: formatSites,
 			onError: () =>
 				dispatch(
 					errorNotice( translate( 'Failed to retrieve your sites. Please try again later.' ) )
