+++ client/jetpack-cloud/sections/agency-dashboard/sites-overview/site-content/index.jsx
@@ -1,6 +1,7 @@
 import { useTranslate } from 'i18n-calypso';
 import useFetchDashboardSites from 'calypso/data/agency-dashboard/use-fetch-dashboard-sites';
 import SiteTable from '../site-table';
+import { formatSites } from '../utils';
 
 import './style.scss';
 
@@ -8,6 +9,8 @@ const SiteContent = () => {
 	const translate = useTranslate();
 
 	const { data, error, isLoading } = useFetchDashboardSites();
+
+	const sites = formatSites( data );
 
 	const columns = {
 		site: translate( 'Site' ),
@@ -17,7 +20,7 @@ const SiteContent = () => {
 		plugin: translate( 'Plugin Updates' ),
 	};
 
-	if ( ! isLoading && ! error && ! data.length ) {
+	if ( ! isLoading && ! error && ! sites.length ) {
 		return <div className="site-content__no-sites">{ translate( 'No active sites' ) }</div>;
 	}
 	return (
@@ -25,7 +28,7 @@ const SiteContent = () => {
 			<SiteTable
 				isFetching={ isLoading }
 				columns={ columns }
-				sites={ data }
+				sites={ sites }
 				isFetchingFailed={ error }
 			/>
 		</>
